Debugger Enablement (#19)
* Debugger Enablement: Switching CGO/gst utility functions from inline to concrete symbols

Unfortunately debug Go builds during linkage due to be unable to find the inline functions.  This commit converts the inline functions to concrete symbols.

The performance implications are untested.

* Debugger Enablement: gst-base: Switching CGO/gst utility functions from inline to concrete symbols

Unfortunately debug Go builds during linkage due to be unable to find the inline functions.  This commit converts the inline functions to concrete symbols.

The performance implications are untested.

enable debugger for base

* Debugger Enablement: gst-app: Switching CGO/gst utility functions from inline to concrete symbols

Unfortunately debug Go builds during linkage due to be unable to find the inline functions.  This commit converts the inline functions to concrete symbols.

The performance implications are untested.

* Debugger Enablement: Removing clashes with go-glib