webrtc: fix nil deref marshal *SessionDescription
When the SessionDescription pointer is nil, ToGValue() will crash
dereferencing the pointer. Fix this by avoiding the dereference.

The ability to marshal a NULL value is required for compatibility with
other languages. This change has been tested with the custom signalling
from gst-plugins-rs, in which a signaller may emit a `session-requested`
signal with an optional offer SDP.